+++ classificacao/Code Feature Extraction/varianceFeature.m
@@ -11,7 +11,7 @@ function [feature] = varianceFeature(X, ~)
     for i = tam/2+1:h-tam/2
 		for j = tam/2+1:w-tam/2
 			aux = gray(i-tam/2:i+tam/2, j-tam/2:j+tam/2);
-			feature = [feature var(aux)];
+			feature = [feature var(double(aux))];
        	end
     end
 end
